Move profiling functions into separate folder
